Former-commit-id: 3e01960b8441d924e11c6e2b45e401b0237d6ea3

+++ api/c++/cub.cc
@@ -23,7 +23,7 @@ CUB::~CUB()
 //------------------------------------------------------------
 void CUB::headin()
 {
-    int nn[SF_MAX_DIM];
+    off_t nn[SF_MAX_DIM];
     char key[3];
 
     nd = sf_filedims(file_,nn);
